LU-2139 ofd: Do async commit if SOFT_SYNC is seen
This change implements the server side portion of the OBD_BRW_SOFT_SYNC
agreement. When this flag is seen on an incoming BRW page, the server
will begin an asynchronous commit operation via dt_commit_async.

The intention is for this operation to commit pages belonging to a
client which has "too many" outstanding unstable pages in its cache.
Upon the commit operation completing, the last_commited value should
percolate out to the client of interest, allowing it to drop some of
its pinned unstable pages.

Unfortunately, there is currently no optimization to proactively send
the last_committed value to the client immediately after the commit
operation succeeds. It must be piggybacked on any future IO or
pings made by the client.

Added fed_soft_sync_count which is set to a limit of 16 SOFT_SYNC
RPCs, and it will be reset as soon as there is a non-SOFT_SYNC RPC.

Added ofd_soft_sync_limit as a tunable proc value for the limit of
fed_soft_sync_count.

This patch is back-ported from the following one:
Lustre-commit: 9761d5c52aeef31a8c3112c3fb3ec9e24b37c800
Lustre-change: http://review.whamcloud.com/5935

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Ic7cf4b8267d0f4b35b80f82fe1dd9f7eed64c8f8
Signed-off-by: Lai Siyao <lai.siyao@intel.com>