LU-5474 tests: sanity-hsm check status using hsm_state
Many tests in sanity-hsm look at hsm/actions on
the MDT to check that archive commands completed. For test_90,
we now check the hsm_state of the archived/restored file
on the client.

Test 90 now uses a local disk for archive to speed up the
archive requests. Archiving to a local drive is used
to speed up archiving for sanity-hsm test 40.

Test 90 is removed from the exception list.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I0beee30b681d4b80f23d33cb42ff5b2944fc21d1