LU-5953 build: use installed OFED by default
During LNET autoconf phase, if OFED installed and its headers
are available, default to use it instead of in-Kernel IB driver.

Presently to do so, browse known default install paths for
headers of all OFED versions.
A more sophisticated detection mechanism may be required to
find if OFED modules are loaded, from which version/RPMs,
from/with which location/prefix, and if corresponding devel
headers are available.

This is intended to work for Lustre build from source and from DKMS.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: I82639f8392d5fe707a3b1a1719d53ab937e918b5