LU-5474 tests: sanity-hsm test_90 use local HSM_ARCHIVE
sanity-hsm test 90 suffers from frequent failures due to
slow archive speeds. If the existing arechive is not
local, test 90 now uses a local disk for archive to speed
the archive process.

sanity-hsm test 40 was modified to look on the SINGLEAGT
node to check if the archive is a local disk for
HSM_ARCHIVE.

copytool_cleanup was modified to match copytool_setup;
remove the contents of $hsm_root and not $hsm_root itself.

Test 90 is removed from the exception list.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I0beee30b681d4b80f23d33cb42ff5b2944fc21d1