LU-5935 lnet: Fixes to make lnetctl function as expected.
During testing of the lnetctl utulity I ran into some
issues. One of the problems is when you print out help
for peer_credits you get help for stats. The second
problem was for grabbing stats data. No data was returned
but itstead it reported a error. The reason for this
is that libcfs_ioctl_getdata test to see if the size of
the data passed in is less than struct libcfs_ioctl_data
in size. For the stats function its data structure
struct lnet_ioctl_lnet_stats is smaller than what is
allowed. This patch fixes the help option for peer_credits
to return the proper help message and enables proper
reporting of stats data. For the stats function we fake
the size to be equal to libcfs_ioctl_getdata.

Change-Id: Ib5abc288d94b0f93549312a3fe283cbb8205febf
Signed-off-by: James Simmons <uja.ornl@gmail.com>