LU-5935 lnet: Fixes to make lnetctl function as expected.
During testing of the lnetctl utility I ran into some
issues. One of the problems is when you print out help
for peer_credits you get help for stats. The second
problem was for grabbing stats data. No data was returned
but instead it reported a error. The reason for this
is that libcfs_ioctl_getdata test to see if the size of
the data passed in is less than struct libcfs_ioctl_data
in size. For the stats function its data structure
struct lnet_ioctl_lnet_stats is smaller than what is
allowed. We remove the size check since the worry is
really to large of buffer being passed in. The help
option for peer_credits is set to return the proper help
message.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ib5abc288d94b0f93549312a3fe283cbb8205febf