LU-5781 osc: check for extents under lock for lru cancel
osc_lock_weight() should try to cancel PW locks as well, however
checking for pages in gang_lookup as was done in LU-3321 wasted
most of the time for looking through cached unused pages.

replace it by a check for osc_extents, which covers regions of
pages and exists only till the IO completion.

Signed-off-by: Vitaly Fertman <vitaly_fertman@xyratex.com>
Change-Id: Id3f93ac0b1e3643269b32376c75e1f3a67715430
Seagate-bug-id: MRP-2176