LU-5555 lprocfs: expire jobstats after /proc access
Currently, jobstats for old jobids are only expired when a new jobid
arrives.  If no new jobstats RPCs have arrived in a long time, the
old jobstats will not be expired even though the /proc/.../job_stats
file is read repeatedly.  This shouldn't be a problem for normal job
submission patterns, but if there is a big gap in new jobid arrival
then the old jobids should be expired.

Expire old jobids after job_stats is read, and move the expiry to the
arrival of new jobids instead of for every RPC to reduce overhead.

Add a test for jobstats arrival.

Test-parameters: fortestonly

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Icfd65ed7cd7beb26a9b566667e2686afe23ebbe5