INTL-157 osd-zfs: Preserve lu_buf when listing EAs
Current osd_xattr_list() in osd-zfs modifies its lu_buf argument,
leaving lu_buf::lb_buf pointing to the end of the original buffer.
This behavior is unnecessary and counter-intuitive.  This patch
preserves the content of the lu_buf argument and updates OSD API to
reduce the chance of future regressions.

Change-Id: I0d3e334848734ca6f44778e87b5ff3b044826784
Signed-off-by: Li Wei <wei.g.li@intel.com>