LU-6027 osd-zfs: Preserve lu_buf when listing EAs
Current osd_xattr_list() in osd-zfs modifies its lu_buf argument,
leaving lu_buf::lb_buf pointing to the end of the original buffer.
This behavior is unnecessary and counter-intuitive.  This patch
preserves the content of the lu_buf argument and updates OSD API to
reduce the chance of future regressions.

Change-Id: Ia17d384c026583ebc7c2e52ce9d46dd2c50fd6ea
Signed-off-by: Li Wei <wei.g.li@intel.com>