LU-5812 osc: remove obd_unstable_pages from osc_announce_cached
The following patch on master branch removed the obd_unstable_pages
tracking:
Lustre-commit: 48eceeec98da3db2c93d980bb5f7e1c1f7c333cc
Lustre-change: http://review.whamcloud.com/10003

However, the codes depend on the patches for LU-3321 that are not in
Lustre b2_5 branch. So, this patch only removes the obd_unstable_pages
tracking from osc_announce_cached() so as to quiet the error message.

Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I07d558a3fb62a679f3e50779b605719fcdf3e0ee