LU-4955 lfsck: reset param when the LFSCK is reset
If "--reset" or "-r" is specified when start the LFSCK manually,
then all non-specified parameters will be set as default values.

If the former LFSCK instance completed, then when re-trigger the
LFSCK, it will restart from the beginning, and all non-specified
parameters will be set as default values.

It is back port to b2_5:
Lustre-change: http://review.whamcloud.com/10091
Lustre-commit: cb22837bcded8f95461c0d4760b2b9add0956e71

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I8f1b08a8dd003ed68cfd0c36340708ad8a904219