LU-5275 lprocfs: mimick new kernels procfs functions
Lustre with newer kernels uses a few functions which don't
exist in older distro's kernels. Here we take the working
lustre proc handling functions and rework them to mimick
the new kernel functions to limit the amount of changes
between the Intel Lustre branch and upstream Lustre in the
lastest kernels.

Change-Id: Ieff19f0216770da94f29562d51abcbf5869bad34
Signed-off-by: James Simmons <uja.ornl@gmail.com>