LU-6010 lnet: prevent assert on LNet module unload
There is a use case where lnet can be unloaded while there are
no NIs configured.  Removing lnet in this case will cause
LNetFini() to be called without a prior call to LNetNIFini().
This will cause the LASSERT(the_lnet.ln_refcount == 0) to be
triggered.

To deal with this use case, LNetFini() checks if ln_refcount == 1
and calls LNetNIFini().  ln_refcount must needs be 1 in this case
since if other modules have initialized LNet they have no way to
indicate not to configure at least a default NI.  In this case
if lustre_rmmod is called it will not be able to unload lnet due
to lnd modules still loaded.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I0f283eeb395fa9a076a4d65ab3edd5e7807fc169