LU-6010 lnet: prevent assert on LNet module unload
There is a use case where lnet can be unloaded while there are
no NIs configured.  Removing lnet in this case will cause
LNetFini() to be called without a prior call to LNetNIFini().
This will cause the LASSERT(the_lnet.ln_refcount == 0) to be
triggered.

To deal with this use case, in fini_lnet() check if the number
of NIs is exactly 1.  If so then that means the LOLND is the only
LND currently loaded, and we will need to call LNetNIFini().

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I0f283eeb395fa9a076a4d65ab3edd5e7807fc169