LU-3538 dne: eliminate dependency in striped dir creation
DNE needs to eliminate dependency between 'mkdir a/b' and any
operation on 'a' because 'b' may be a striped directory, which is
able to recover from update log, if 'a' can't be recovered, it
will cause problem.

To achieve this, 'mkdir a/b' will fake to be a distributed
transaction to make sure previous modification on 'a' is committed
to disk.

And mkdir also needs to save child lock, so that subsequent mkdir
under it can know the confliction.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I1887f7c7978d94695563cf77290148b9c35170c7