LU-3538 ldlm: add strict COS lock
DNE needs to eliminate dependency between 'mkdir a' and 'mkdir a/b',
because 'b' may be a striped directory, which is able to recover
from mirrored MDS, if 'a' can't be recovered, it will cause problem.

To achieve this, 'mkdir a/b' will enqueue strict lock on 'a', which
means this lock will conflict with lock even from the same client.

And mkdir also needs to save child lock, so that subsequent mkdir
under it can know the confliction.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I1887f7c7978d94695563cf77290148b9c35170c7