LU-6055 hsm: improve error handling in sanity-hsm
In path2fid() propagate errors from 'lfs path2fid'.  In
file_creation_failure() redirect the output of df to stderr so that it
is not captured by command substitution.

This is a backport to b2_5 of
Lustre change: http://review.whamcloud.com/13154
Lustre commit: 27f65af06d0906856041cbcf2537cfabe2a90583

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I7fabf46af4aa36c6affef45cf8c9b044a7aa29dd
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>