LU-6025 lctl: remove bad error checking from nodemap_cmd()
There was some error checking that didn't really make sense in
nodemap_cmd() and was causing false errors. memcpy() doesn't have an
error condition, so there's no need to check it.

This patch also includes an update to the lctl man page for the
nodemap functions, courtesy of Justin Miller of Indiana U.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I58225f8ebb1d8dc0941534503f939b213d57c27f