LU-6026 tests: modify sanity-sec to flush MDT locks
In order to create test directories, the test enables admin mode on
a client, and then disables it. It's possible that the client could
then start the nodemap test before the MDT has flushed all its locks.
This patch flushes the locks on the client after re-disabling admin
mode, so that the proper mapping is guaranteed.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: If0707cfc3e93b25538c11347d9d7a1178fcc151f