LU-4931 ladvise: Add feature of giving file access advices
Linux kernel provide mechanism to applications to give advices
or hints about how to choose appropriate read-ahead and caching
techniques to the corresponding files. However, this is only a
client side mechanism which is not enough for distributed file
systems like Lustre, because in order to tune system-wide
cache or read-ahead policies, server needs to understand the
advices too.

This patch adds a new feature which provides new APIs and utils
to give advices about the access pattern of Lustre file with the
purpose of performance improvement. It can pass advices directly
through Lustre client to server/storage side.

Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: I6de14ed87199949963360d393897ff79826c9819