LU-6026 tests: modify sanity-sec use right LNET, flush locks
There were a couple of spots in sanity-sec that had the @tcp LNET type
hardcoded in. This caused it to fail on other LNET networks, like
o2ib. This patch uses the $NETTYPES variable to determine the correct
NID.

There is also a chance that the client will not flush its locks prior
to doing the nodemap tests. This patch adds a line to force a flush.

Test-Parameters: alwaysuploadlogs nettypes=o2ib testlist=sanity-sec
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I68c6400217100939c171ad3707ed834c98a7755a