LU-6081 hsm: adding migration support
Currently file migration is done by "lfs migrate". This has a couple
problems. It doesn't scale well to large files and/or large number of
files to copy. If an external tools (such as Robinhood) wants to
migrate a file, it has to spawn lfs to do the job, for every file,
which has a big overhead, or copy the file itself.

A solution to this issue is to move the copy operation into the HSM
copytool. The copytool already knows how to copy files, and Lustre has
control over these files. Migrating a file, from an application point
of view, becomes more or less the same as archiving a file. Since the
file copy is done by the copytool, this copy can be optimized in one
place.

This patch adds the ability for HSM to migrate a file within a single
Lustre filesystem. The posix copytool is enhanced to support the file
migration, and the lfs migrate command now supports that option with
-H and -a flags.

Signed-off-by: Patrick Farrell <paf@cray.com>
Signed-off-by: frank zago <fzago@cray.com>
Change-Id: Id72db5acb3f84292b081efac4657520f0054778b