LU-5275 lprocfs: sync names to upstream kernel lustre client
When seq file handling was introduced to lustre we had to
create duplicate functions that handle a different type of
struct lprocfs_vars which was names struct lprocfs_seq_vars.
Now that lustre has moved to using only seq_file we can
rename the special *seq* functions and structures to what is
exactly in the upstream linux kernel. This helps to greatly
reduce the difference between upstream and the Intel branch.

Change-Id: Ic4f7eac105736c691ea4b37438352e5542ce344c
Signed-off-by: James Simmons <uja.ornl@gmail.com>