LU-6098 tests: start debug_daemon before server start
Alter test-framework.sh:setupall() to  start the debug_daemon
before mounting the servers, instead of waiting to start
it just before mounting the client.

This makes examining the server startup process easier.

De-duplicate the host list in do_nodes() to
eliminate the potential race condition created when
commands are run multiple times on a single host.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Change-Id: Ida92923807ce9c6c059376e937555fa1f840e2f3