LU-5969 lustreapi: replace llapi_get_version()
The existing llapi_get_version() helper has an awkward calling
style, requiring a temporary input buffer, as well as an output
pointer.  It also typically returns an unsightly string for
the build, instead of a simple version that would be expected.

Replace this with two separate functions:
llapi_get_build_version() which returns the current ugly build string
llapi_get_release_version() which returns the simple version string

These versions are extracted from /proc/fs/lustre/version instead
of from an ioctl, to avoid the messiness that is Lustre ioctls.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I681a353e6f65fd1ec2d1434fe38dfa55ed3ebbe5