LU-6148 ldlm: Strided locking prototype
This is a prototype/proof of concept, not intended to land
as is.  It is missing at least one major set of changes
for handling conflicting locks, and has a great deal of
debug code which is intended as development help rather
than intended to land.  This is intended to start
concept and design review, and if that goes well, it will
be cleaned up.

This adds a new LDLM layer feature, strided locking.
Strided locking makes it possible to improve shared file IO
performance.  Currently, if more than one client tries to
write to the portion of a file on a given OST (IE, >1
client per OST), the two clients have lock contention
problems, and are serialized, even if their writes are
non-overlapping.  The usual case for this is strided IO,
where clients write alternating blocks.

If a group of clients commit want to do strided IO to a
single file, they can use an ioctl to request strided
locking behavior.

The client provides the number of clients in the strided
IO, and when that client does an IO request, the lock
granted is a periodic lock which is periodic in the
number of clients.  Locks are granted only on stripe size
segments.

IE, if there are two clients in the IO, a lock which
covers every 2nd block is issued to each client, starting
from the block where the IO was done.

This allows multiple clients to write to the same file on
a particular OST without being serialized by lock
contention.

Design doc and other details forthcoming in LU-6148.

Test-Parameters: fortestonly
Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: Ia529ac7cb94b4fb8e3d81e6080c16787107b4df0