LU-6122 lnet: Allocate the correct number of rtr buffers
This patch ensures that the correct number of router buffers are
allocated.  It keeps a new count that keeps track of the number of
buffers allocated.  Another count keeps the number of buffers
requested. The number of buffers allocated is set when creating
new buffers and reduced when buffers are freed.

The number of requested buffer is set when the buffers are
allocated and is checked when credits are returned to determine
whether the buffer should be freed or kept.

When the above check is performed a mutex is locked to ensure that
no other buffers are added while the buffer is being freed.  After
the mutex is locked the condition is checked again to deal with
the scenario where the user requests an increase in buffers, that
thread grabs the mutex and starts to allocate buffers, during this
process a buffer is being returned and is about to be freed.  At
this point the condition could be true, but after the buffers
are created, the condition is no longer true.  In order to avoid
this the condition is checked again once the mutex is acquired.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I96627cc8ba3d3d70a0bf581b21ccd3c9b2de327f